Extract active components list into its own Components tab
