+++ .changeset/witty-melons-cheer.md
@@ -1,0 +1,5 @@
+---
+"@osdk/react-devtools": patch
+---
+
+Move the active components list out of the Debugging tab into its own Components tab in the devtools panel
